NexusSagaServerInternal-develop/siro_server/requirements.txt
Flask-SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
Flask==2.0.3
Werkzeug==2.0.3
eventlet
#python-socketio<5.0.0
requests
websocket-client